♻️ Refactor: 자동 로그인 리팩토링
Related to: #193
