+++ app/src/main/AndroidManifest.xml
@@ -36,14 +36,14 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".presentation.ui.splash.OnBoardingActivity"
+            android:name=".presentation.ui.onBoarding.OnBoardingActivity"
             android:exported="false" />
         <activity
             android:name=".presentation.ui.MainActivity"
             android:exported="false"
             android:windowSoftInputMode="adjustPan" />
         <activity
-            android:name=".presentation.ui.splash.SplashActivity"
+            android:name=".presentation.ui.SplashActivity"
             android:exported="true"
             android:theme="@style/Theme.App.Starting">
             <intent-filter>
