+++ app/src/main/java/com/mongmong/namo/data/datasource/auth/RemoteAuthDataSource.kt
@@ -71,16 +71,16 @@ class RemoteAuthDataSource @Inject constructor(
                 refreshToken = ""
             )
         )
-//        withContext(Dispatchers.IO) {
-//            runCatching {
-//                authApiService.refreshToken(tokenBody)
-//            }.onSuccess {
-//                Log.d("RemoteAuthDataSource", "postNaverLogin Success $it")
-//                refreshResponse = it
-//            }.onFailure {
-//                Log.d("RemoteAuthDataSource", "postNaverLogin Fail $it")
-//            }
-//        }
+        withContext(Dispatchers.IO) {
+            runCatching {
+                authApiService.refreshToken(tokenBody)
+            }.onSuccess {
+                Log.d("RemoteAuthDataSource", "postNaverLogin Success $it")
+                refreshResponse = it
+            }.onFailure {
+                Log.d("RemoteAuthDataSource", "postNaverLogin Fail $it")
+            }
+        }
         return refreshResponse
     }
