+++ app/src/main/java/com/mongmong/namo/data/remote/AuthApiService.kt
@@ -28,7 +28,7 @@ interface AuthApiService {
     @POST("auths/reissuance")
     fun refreshToken(
         @Body body: TokenBody
-    ): Call<RefreshResponse>
+    ): RefreshResponse
 
     // 로그아웃
     @POST("auths/logout")
