+++ app/src/main/java/com/mongmong/namo/presentation/config/XAccessTokenInterceptor.kt
@@ -1,17 +1,25 @@
 package com.mongmong.namo.presentation.config
 
 import android.util.Log
+import com.mongmong.namo.data.remote.AuthApiService
 import com.mongmong.namo.presentation.config.ApplicationClass.Companion.X_ACCESS_TOKEN
 import com.mongmong.namo.presentation.config.ApplicationClass.Companion.X_REFRESH_TOKEN
 import com.mongmong.namo.presentation.config.ApplicationClass.Companion.sSharedPreferences
-import com.mongmong.namo.data.remote.auth.RefreshService
+import com.mongmong.namo.domain.model.RefreshResponse
 import com.mongmong.namo.domain.model.TokenBody
+import com.mongmong.namo.domain.repositories.AuthRepository
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.runBlocking
 import okhttp3.Interceptor
 import okhttp3.Response
 import java.io.IOException
 import javax.inject.Inject
 
-class XAccessTokenInterceptor @Inject constructor() : Interceptor {
+class XAccessTokenInterceptor @Inject constructor(
+    private val apiService: AuthApiService
+) : Interceptor {
     @Throws(IOException::class)
     override fun intercept(chain: Interceptor.Chain): Response {
         val request = chain.request()
@@ -44,21 +52,18 @@ class XAccessTokenInterceptor @Inject constructor() : Interceptor {
 
                 // 재발급 api 호출
                 // 받은 토큰 다시 넣어서 기존 api 재호출
-                val newTokenResponse = RefreshService().tryTokenRefresh(TokenBody(accessToken.toString(), refreshToken))
-                val responseBody = newTokenResponse.body()
+                val newTokenResponse = runBlocking { apiService.refreshToken(TokenBody(accessToken.toString(), refreshToken)) }
 
                 Log.d("Token", newTokenResponse.toString())
 
-                if (responseBody != null) {
+                if (newTokenResponse != null) {
                     sSharedPreferences.edit()
-                        .putString(X_REFRESH_TOKEN, responseBody.result.refreshToken)
-                        .putString(X_ACCESS_TOKEN, responseBody.result.accessToken)
+                        .putString(X_REFRESH_TOKEN, newTokenResponse.result.refreshToken)
+                        .putString(X_ACCESS_TOKEN, newTokenResponse.result.accessToken)
                         .apply()
                     Log.d("onTokenResponse", "토큰 재발급 성공!")
-//                    Log.d("onTokenResponse", "${newTokenResponse.code()}\n" + "${newTokenResponse.body()!!.result}")
 
-                    //val newJwtToken: String? = sSharedPreferences.getString(X_ACCESS_TOKEN, null)
-                    val newJwtToken = responseBody.result.accessToken
+                    val newJwtToken = newTokenResponse.result.accessToken
 
                     // 새로운 토큰으로 하던 작업 서버 재요청
                     val finalRequest = chain.request().newBuilder()
@@ -69,7 +74,7 @@ class XAccessTokenInterceptor @Inject constructor() : Interceptor {
                     return chain.proceed(finalRequest)
                 }
 
-                if (newTokenResponse.body()?.code == 403) { // 리프레시 토큰 만료
+                if (newTokenResponse.code == 403) { // 리프레시 토큰 만료
                     Log.d("Token", "403 리프레시 토큰 만료")
                     //TODO: 로그인 다시하기
