+++ app/src/main/java/com/mongmong/namo/presentation/ui/onBoarding/OnBoardingFragment.kt
@@ -1,4 +1,4 @@
-package com.mongmong.namo.presentation.ui.splash
+package com.mongmong.namo.presentation.ui.onBoarding
 
 import android.content.Context
 import android.os.Bundle
@@ -6,11 +6,10 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.navigation.Navigation.findNavController
 import androidx.navigation.fragment.findNavController
 import com.mongmong.namo.R
 import com.mongmong.namo.databinding.FragmentOnboardingBinding
-import com.mongmong.namo.presentation.ui.splash.adapter.OnboardingVPAdapter
+import com.mongmong.namo.presentation.ui.onBoarding.adapter.OnboardingVPAdapter
 import com.tbuonomo.viewpagerdotsindicator.DotsIndicator
 
 class OnBoardingFragment : Fragment() {
