--- app/src/main/java/com/mongmong/namo/presentation/ui/splash/SplashActivity.kt
@@ -1,72 +1,0 @@
-package com.mongmong.namo.presentation.ui.splash
-
-import android.content.Intent
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.core.splashscreen.SplashScreen
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.mongmong.namo.data.remote.auth.RefreshService
-import com.mongmong.namo.data.remote.auth.SplashView
-import com.mongmong.namo.databinding.ActivitySplashBinding
-import com.mongmong.namo.domain.model.LoginResponse
-import com.mongmong.namo.domain.model.RefreshResponse
-import com.mongmong.namo.domain.model.TokenBody
-import com.mongmong.namo.presentation.config.ApplicationClass
-import com.mongmong.namo.presentation.ui.MainActivity
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.launch
-
-class SplashActivity : AppCompatActivity(), SplashView {
-
-    private lateinit var binding : ActivitySplashBinding
-    private lateinit var splashScreen: SplashScreen
-    private val isDataLoaded = MutableStateFlow(false)
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        splashScreen = installSplashScreen()
-        splashScreen.setKeepOnScreenCondition {
-            !isDataLoaded.value
-        }
-
-        autoLogin()
-
-        binding = ActivitySplashBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-    }
-
-    private fun autoLogin() {
-        CoroutineScope(Dispatchers.IO).launch {
-            val splashService = RefreshService()
-            splashService.setSplashView(this@SplashActivity)
-
-            val accessToken = ApplicationClass.sSharedPreferences.getString(ApplicationClass.X_ACCESS_TOKEN, null)
-            val refreshToken = ApplicationClass.sSharedPreferences.getString(ApplicationClass.X_REFRESH_TOKEN, null)
-            splashService.splashTokenRefresh(TokenBody(accessToken.toString(), refreshToken.toString()))
-        }
-    }
-
-    override fun onVerifyTokenSuccess(response: RefreshResponse) {
-        ApplicationClass.sSharedPreferences.edit()
-            .putString(ApplicationClass.X_REFRESH_TOKEN, response.result.refreshToken)
-            .putString(ApplicationClass.X_ACCESS_TOKEN, response.result.accessToken)
-            .apply()
-
-        runOnUiThread() {
-            startActivity(Intent(this@SplashActivity, MainActivity::class.java))
-            finish()
-            isDataLoaded.value = true
-        }
-    }
-
-    override fun onVerifyTokenFailure(message: String) {
-        runOnUiThread() {
-            startActivity(Intent(this@SplashActivity, OnBoardingActivity::class.java))
-            finish()
-            overridePendingTransition(0, 0)
-            isDataLoaded.value = true
-        }
-    }
-}
