--- app/src/main/java/com/mongmong/namo/presentation/utils/Constants.kt
@@ -1,7 +1,0 @@
-package com.mongmong.namo.presentation.utils
-
-object Constants {
-    /** 약관 관련 */
-    const val TERM_URL = "https://mynamo.notion.site/fa0fbc3ae51a44679194cbd129243573?pvs=4" // 이용 약관
-    const val POLICY_URL = "https://mynamo.notion.site/26b1f16363f14517b2aaf667b0c2b807?pvs=4" // 개인정보 처리방침
-}
