+++ app/src/main/res/navigation/nav_splash.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/onBoardingFragment"
-        android:name="com.mongmong.namo.presentation.ui.splash.OnBoardingFragment"
+        android:name="com.mongmong.namo.presentation.ui.onBoarding.OnBoardingFragment"
         android:label="OnBoardingFragment"
         tools:layout="@layout/fragment_onboarding">
         <action
